requirements.txt
gym
dm-haiku
optax
rlax
pytest